Move ECS, SSH, and Secrets services to Engine
